+++ tests/Neo.Compiler.MSIL.UnitTests/UnitTest_Delegate.cs
@@ -1,5 +1,11 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Neo.Compiler.MSIL.Extensions;
 using Neo.Compiler.MSIL.UnitTests.Utils;
+using Neo.IO;
+using Neo.Ledger;
+using Neo.Persistence;
+using Neo.VM;
+using Neo.VM.Types;
 
 namespace Neo.Compiler.MSIL.UnitTests
 {
@@ -7,11 +13,14 @@ namespace Neo.Compiler.MSIL.UnitTests
     public class UnitTest_Delegate
     {
         private TestEngine testengine;
+        private readonly TestSnapshot snapshot = new TestSnapshot();
 
         [TestInitialize]
         public void Init()
         {
-            testengine = new TestEngine();
+            snapshot.DeployNativeContracts();
+
+            testengine = new TestEngine(snapshot: snapshot);
             testengine.AddEntryScript("./TestClasses/Contract_Delegate.cs");
         }
 
@@ -21,6 +30,21 @@ namespace Neo.Compiler.MSIL.UnitTests
             testengine.Reset();
             var result = testengine.ExecuteTestCaseStandard("sumFunc", 2, 3).Pop();
             Assert.AreEqual(5, result.GetInteger());
+        }
+
+        [TestMethod]
+        public void TestDelegateCall()
+        {
+            var token = UInt160.Parse("0x4961bf0ab79370b23dc45cde29f568d0e0fa6e93"); // NEO token
+
+            testengine.Reset();
+            var result = testengine.ExecuteTestCaseStandard("testDynamicCall", token.ToArray(), "symbol");
+            Assert.AreEqual(VMState.HALT, testengine.State);
+            Assert.AreEqual(1, result.Count);
+
+            var item = result.Pop();
+            Assert.IsInstanceOfType(item, typeof(ByteString));
+            Assert.AreEqual("NEO", item.GetString());
         }
     }
 }
